make ts code runnable
